refactor viewmodel into separate view models, 1 per fragment
