+++ PennMobile/src/main/java/com/pennapps/labs/pennmobile/gsr/viewmodels/BookGsrViewModel.kt
@@ -11,7 +11,7 @@ import kotlinx.coroutines.launch
 import javax.inject.Inject
 
 @HiltViewModel
-class GsrViewModel
+class BookGsrViewModel
     @Inject
     constructor(
         private val repository: GsrRepo,
@@ -24,12 +24,6 @@ class GsrViewModel
 
         private val _bookingSuccess = MutableStateFlow(false)
         val bookingSuccess = _bookingSuccess.asStateFlow()
-
-        private val _isCancelling = MutableStateFlow(false)
-        val isCancelling = _isCancelling.asStateFlow()
-
-        private val _cancelSuccess = MutableStateFlow(false)
-        val cancelSuccess = _cancelSuccess.asStateFlow()
 
         val savedUserInfo = repository.getSavedUserInfo()
 
@@ -56,34 +50,11 @@ class GsrViewModel
                     repository.bookGsr(startTime, endTime, gid, roomId, roomName, firstName, lastName, email)
                     _bookingSuccess.value = true
                 } catch (e: Exception) {
-                    Log.e("GsrViewModel", "Booking failed", e)
+                    Log.e("BookGsrViewModel", "Booking failed", e)
                     _error.value = e
                 } finally {
                     _isBooking.value = false
                 }
             }
-        }
-
-        fun cancelGsr(
-            bookingId: String?,
-            sessionId: String?,
-        ) {
-            viewModelScope.launch {
-                _isCancelling.value = true
-                _error.value = null
-                try {
-                    repository.cancelGsr(bookingId, sessionId)
-                    _cancelSuccess.value = true
-                } catch (e: Exception) {
-                    Log.e("GsrViewModel", "Cancellation failed", e)
-                    _error.value = e
-                } finally {
-                    _isCancelling.value = false
-                }
-            }
-        }
-
-        fun resetCancelSuccess() {
-            _cancelSuccess.value = false
         }
     }
